Flash/Pause/Sprint
Hope Nothing messes up this time...I tried to limit it as best I could.
The only things that worry me now are the .log, the .json and the .ini
however I didn't even see those in the danger list that Noy said so I'm
just gonna hope this doens't destroy anything.