Organization/Checking Commit
Organized Assets into proper folders. The Lighting system that I had set
up previously broke due to reset/commit editting issues. If one of you
can pick that up that'd be great. Also broke the link for the Normal
mapping for the Box Crates cause it was a bad normal with splotches
every time you shined light on it.