Zombie & Music Content Commit
Two Fully Textured & Animated Zombies ready for Programming. The music
is something I found for the Main menu I'm thinking of doing a short
1:30 or so video that'll Loop. It'll show a cinematic shot of a "Paused"
zombie mob heading towards the character which which will have the Gun
firing effect at the end of the gun. Each of these scenes will be
panning shots. (There is obviously a lot more to describe but I don't
want to bog this commit too much).