Ammo HUD & Zombie Edit
Ammo works great now, most likely would need some tweaking with amounts
and firerate however it works fine for right now. There is an Ammobox
which when stepped on will restore the Max Ammo by 40 (which is a whole
clip, not the final number might be 120). Also editted Zombie1 (the
taller zombie on the right) so that when "killed" will not replay
animation set and will boolean as false for bAlive.